Endpointer function

The epR function is a tool for the identification of humane endpoints using single outcome variables from laboratory animal experiments. Originally, it was developed for using body weight but as values are normalized other continous variables are suited as well. The algorithm is highly functional in identifying strong deviations from a windowed normality. The hypothesis behind this is: larger deviations always point to severity.

epR(td = td, org = FALSE, wl = 6, SDwdth = 2, mad = FALSE,
  ltype = "b", dotcolor = "black", uprcol = "darkgreen",
  lwrcol = "magenta", cex = 1, cex.axis = 1, cex.lab = 1,
  xlim = NULL, ylim = NULL, pch = 19, blind = FALSE,
  ignupr = FALSE, xlab = "time", ylab = "Moving average (%)",
  main = NULL)

Arguments

td
testdata data.frame with n unique rows and p subsequent time points (e.g. days)

org
boolean (TRUE/FALSE) for using original values. If FALSE, data are normalized.

wl
SD window length (default is 6)

SDwdth
width of the standard deviation around the moving average (default is 2.5)

mad
boolean - use mean absolute deviation as quasi-clinical scoring constraint (default FALSE)

ltype
line type in shown plot

dotcolor
color of the shown dots (default "black")

uprcol
color of the upper threshold violation indicators

lwrcol
color of the upper threshold violation indicators

cex
point size

cex.axis
axis tick size

cex.lab
label size

xlim
range of x-axis (if set to NULL (default), plot will adapt automatically to given range - may not be nice!)

ylim
range of <-axis (if set to NULL (default), plot will adapt automatically to given range - may not be nice!)

pch
R-specific plot symbol for shown dots (default is 19)

blind
boolean (TRUE/FALSE) - if set to TRUE, no plot will be shown (default is FALSE)

ignupr
boolean (TRUE/FALSE) - ignore upper threshold violations (default is FALSE)

xlab
x-axis label

ylab
y-axis label

main
title

Value

data.frame with enpointeR results (n=number of data points, timepoint=index of marked endpoint, where=upper or lower boundary)

Examples

epR(as.numeric(gliodat[1,3:length(gliodat[1,])]), blind=TRUE )

#>    n timepoint value where
#> 1 29        12 99.25 lower
#> 2 29        15 94.14 lower