Animal Selection for endpointR Analysis

Using the testdata output from ep_load, the ep-select function subsets the testdata by a user-defined unique animal index. If the input is numeric, e.g. n=1 the first animal in testdata will be chosen. If the input is a character string, e.g. "ZW21", this specific animal from the testdata will be chosen. Animal ids should not be numeric.

ep_select(testdata, n = 1)

Arguments

testdata
data.frame with test data as provided by ep_load

n
animal index (numeric or character (animal id))

Value

vector of a single animal for distinct time points.

Examples

ep_select(gliodat, n = 1)

#>  [1] 328 324 319 326 315 328 331 331 328 334 336 329 330 326 308