Rat Body Weight Data Glioblastoma Model

Rat body weight data from a intracranial Glioblastoma model. 63 unique animals in three treatment groups (Control, Microinjection and Resection) were monitored on a daily basis. The general model indicates that animal well-being is not substantially impaired until a certain time-point at around 12-16 days is reached. Here, most animals decompensate rapidly and had to be euthanized because the defined humane endpoint for the study was reached. The maximum follow-up is 28 days, although most animals had to be euthanized well before that. In the resection group day 8 is the day of operation. These animals survive longer but a small drop in body weight can be seen after day 8, together with rapid recovery.

data(gliodat)

Format

A 63 x 31 data frame with post-op data.

Examples

data(gliodat)
head(gliodat)

#>   Animal_No      Treatment  d0  d1  d2  d3  d4  d5  d6  d7  d8  d9 d10 d11 d12
#> 1       WZ1 Microinjection 328 324 319 326 315 328 331 331 328 334 336 329 330
#> 2      WZ11 Microinjection 397 391 380 379 372 373 374 378 381 381 379 379 376
#> 3      WZ12 Microinjection 415 412 408 402 399 400 400 401 402 405 404 402 402
#> 4      WZ13 Microinjection 414 418 415 412 392 395 398 401 399 400 393 396 392
#> 5      WZ15        Control 417 422 415 406 402 400 406 404 402 390 392 391 387
#> 6      WZ16      Resection 442 430 429 410 417 418 417 417 418 406 405 403 407
#>   d13 d14 d15 d16 d17 d18 d19 d20 d21 d22 d23 d24 d25 d26 d27 d28
#> 1 326 308  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA
#> 2 380 378 350  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA
#> 3 405 410 405 385  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA
#> 4 390 380  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA
#> 5 395 390 391 394 395 397 394 398 398 403 400 401 389 385 362  NA
#> 6 407 410 402 409 409 410 412 412 413 412 385  NA  NA  NA  NA  NA